refactor: regroup everything related to the external configuration domain
